Updated the courses_table logic
When we would add/drop a course, only the user_classes table would get the enrollment update but not the courses_table so that was fixed here. Still having issues with the "FULL" logic whenever the enrollment is the same as the capacity but for now it updates the courses table as well.