Updated the user_classes table display
Before the table would display the data last stored in the local storage that was saved when a class was added/dropped. However, if we logged in with another user, the table being displayed is for the last user logged in but now it loads the table based on the current users data stored in the database.