Updated class enrollment updating
Before, the class enrollment would just update for the current user but was not updating for multiple users. For example, say Math 131 starts as 1/10 enrollment. User 1 adds the course and it should update to 2/10. Log out and log in with user 2 and add the same course, thus, updating to 3/10 and so on until the course is "FULL". Also there was an error issue with the logic when trying to drop a "FULL" course but it was just a wrong condition placement inside the function.